+++ Practica3_cholesky/chol_interface.f95
@@ -2,8 +2,8 @@ module chol_interface
   interface
     subroutine chol(a, deter)
       implicit none
-      real, dimension(:,:), intent(inout) :: a
-      real, intent(out) :: deter
+      real(8), dimension(:,:), intent(inout) :: a
+      real(8), intent(out) :: deter
     end subroutine
   end interface
 end module
