+++ Practica3_cholesky/sistl_interface.f95
@@ -1,9 +1,9 @@
 module sistl_interface
   interface
     subroutine sistl(a, b, w)
-      real, dimension(:,:), intent(in) :: a
-      real, dimension(:), intent(in) :: b
-      real, dimension(:), intent(out) :: w
+      real(8), dimension(:,:), intent(in) :: a
+      real(8), dimension(:), intent(in) :: b
+      real(8), dimension(:), intent(out) :: w
     end subroutine
   end interface
 end module
