+++ Practica3_cholesky/sistu_interface.f95
@@ -1,9 +1,9 @@
 module sistu_interface
   interface
     subroutine sistu(a, b, u)
-      real, dimension(:,:), intent(in) :: a
-      real, dimension(:), intent(in) :: b
-      real, dimension(:), intent(out) :: u
+      real(8), dimension(:,:), intent(in) :: a
+      real(8), dimension(:), intent(in) :: b
+      real(8), dimension(:), intent(out) :: u
     end subroutine
   end interface
 end module
