+++ Practica3_lutri/residuotri.f95
@@ -4,8 +4,8 @@ subroutine residuotri(ad, al, au, u, b, r)
 
   implicit none
 
-  real, dimension(:), intent(in) :: ad, al, au, u, b
-  real, dimension(:), intent(out) :: r
+  real(8), dimension(:), intent(in) :: ad, al, au, u, b
+  real(8), dimension(:), intent(out) :: r
 
   integer :: n
