+++ Practica3_lutri/residuotri_interface.f95
@@ -1,8 +1,8 @@
 module residuotri_interface
   interface
     subroutine residuotri(ad, al, au, u, b, r)
-      real, dimension(:), intent(in) :: ad, al, au, u, b
-      real, dimension(:), intent(out) :: r
+      real(8), dimension(:), intent(in) :: ad, al, au, u, b
+      real(8), dimension(:), intent(out) :: r
     end subroutine
   end interface
 end module
