+++ Practica3_lutri/tridiagxvector_interface.f95
@@ -1,8 +1,8 @@
 module tridiagxvector_interface
   interface
     subroutine tridiagxvector(ad, al, au, v, w)
-      real, dimension(:), intent(in) :: ad, al, au, v
-      real, dimension(:), intent(out) :: w
+      real(8), dimension(:), intent(in) :: ad, al, au, v
+      real(8), dimension(:), intent(out) :: w
     end subroutine
   end interface
 end module
